requirements.txt
customtkinter
Pillow